image_captioning_project/requirements.txt

torch
torchvision
nltk
pandas
Pillow
tqdm

numpy>=1.21.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
transformers>=4.30.0
jupyter>=1.0.0
matplotlib>=3.5.0
tqdm>=4.65.0 